requirements.txt
langchain==0.3.*
langgraph==0.3.*
langchain.groq==0.3.*
streamlit==1.36.*
python-dotenv==1.0.*